% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadnse.r
\name{downloadnse}
\alias{downloadnse}
\title{downloadnse function
The function retrieves the archived data, downloads new data and combine both of the datasets. Then archive the updated version of the data}
\usage{
downloadnse()
}
\description{
downloadnse function
The function retrieves the archived data, downloads new data and combine both of the datasets. Then archive the updated version of the data
}
\keyword{(havingIP)}
\keyword{bourse}
\keyword{downloadnse}
\keyword{kenya}
\keyword{kenya,}
\keyword{market,}
\keyword{nse}
\keyword{stock}
